+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Unreleased
 ----------
 - Removed custom event variants & infrastructure
+- Removed dedicated `UnhandledEvent` type
 - Adjusted event handlers to return a single event only
   - Removed event chaining logic